Have updated the code and readme based on PR comments
