--- _infra/helm/locust/locustfiles/1001_2212_0001.xlsx
@@ -1,3 +1,0 @@
-US006: Load SEFT Collection Instruments,,,,
-Formtype 0002,,,,
-Survey 141,,,,
