--- _infra/helm/locust/locustfiles/requests_seft.json
@@ -1,33 +1,0 @@
-{
-    "requests": [
-        {
-            "method": "GET",
-            "url": "/surveys/todo",
-            "expected_response_text": "Click on the survey name to complete your questionnaire"
-        },
-        {
-            "method": "GET",
-            "harvest_url": {
-                "id": "surveyLongName",
-                "link_text": "Annual Survey of Hours and Earnings"
-            },
-            "expected_response_text": "Annual Survey of Hours and Earnings",
-            "response_status": 200,
-            "grouping": "/surveys/access-survey"
-        },
-        {
-            "method": "GET",
-            "response_status": 200,
-            "url": "/surveys/download-survey",
-            "expected_content_type": "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet",
-            "expected_content_length": "1",
-            "expected_content_disposition": "065_201803_0001.xlsx"
-        },
-        {
-            "method": "POST",
-            "response_status": 200,
-            "url": "/surveys/upload-survey",
-            "expected_response_text": "File uploaded successfully"
-        }
-    ]
-}
