+++ _infra/helm/locust/values.yaml
@@ -2,7 +2,7 @@ locust:
   loadtest:
     name: performance-test
     headless: true
-    locust_locustfile: locustfile_download_upload.py
+    locust_locustfile: locustfile.py
     locust_locustfile_configmap: locust-worker
     locust_host: http://frontstage.performance.svc.cluster.local:9000
     pip_packages:
@@ -19,7 +19,7 @@ locust:
       security_user_name: admin
       security_user_password: secret
       test_respondent_password: password
-      test_respondents: 1
+      test_respondents: 10
       GOOGLE_CLOUD_PROJECT: "ras-rm-performance-20220908"
       GCS_BUCKET_NAME: "ras-rm-performance-20220908-locust"
       CSRF_ENABLED: true
